added typescript compiler and rebuilt Money object
