+++ package.json
@@ -25,6 +25,7 @@
     "superagent": "^1.1.0"
   },
   "devDependencies": {
+    "awesome-typescript-loader": "^0.5.4",
     "babel": "^5.0.8",
     "babel-core": "^5.0.8",
     "babel-loader": "^4.1.0",
@@ -39,6 +40,7 @@
     "gulp-uglify": "^1.0.2",
     "gulp-util": "^3.0.1",
     "json-loader": "^0.5.1",
+    "jsx-typescript": "^1.5.0-alpha.5",
     "map-stream": "0.0.5",
     "mocha": "^2.2.1",
     "mocha-lcov-reporter": "0.0.2",
@@ -48,6 +50,7 @@
     "sinon-chai": "^2.7.0",
     "style-loader": "^0.8.3",
     "stylus-loader": "^1.0.0",
+    "typescript-loader": "^1.1.3",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.0.0",
     "webpack": "^1.7.3",
