+++ tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+        "noImplicitAny": true,
+        "target": "es5",
+        "module": "commonjs"
+    }
+}
